# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: beelms API
  version: 0.1.0
  license:
    name: UNLICENSED
    url: https://example.com/unlicense
  description: |
    OpenAPI спецификация за основните MVP услуги на beelms.
servers:
  - url: https://api.beelms.com/api
    description: Production (примерен домейн, с /api префикс)
  - url: http://127.0.0.1:3000/api
    description: Local development (с /api префикс)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
    AuthToken:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        role:
          type: string
          description: Роля на потребителя (напр. "user", "admin", "teacher", "author", "monitoring")
        emailChangeLimitReached:
          type: boolean
          description: Дали е достигнат лимитът за потвърждения на смяна на email в текущия прозорец
        emailChangeLimitResetAt:
          type: string
          format: date-time
          nullable: true
          description: ISO дата/час, когато лимитът за смяна на email се нулира (ако е приложимо)
...
    AdminUserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          description: Роля на потребителя (напр. "user", "admin", "monitoring", "teacher", "author")
          enum: [user, admin, monitoring, teacher, author]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminUpdateUserInput:
      type: object
      properties:
        active:
          type: boolean
        role:
          type: string
          enum: [user, admin, monitoring, teacher, author]
...
    AdminUsersStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        deactivatedUsers:
          type: integer
        adminUsers:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Auth]
      summary: Регистрация на нов потребител
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        "201":
          description: Успешна регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "400":
          description: Невалидни данни
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Вход на потребител
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: Успешен вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        "401":
          description: Невалидни идентификационни данни
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Auth]
      summary: Заявка за смяна на забравена парола
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Имейл с инструкции е изпратен (ако адресът съществува)
        "400":
          description: Невалидни данни
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Auth]
      summary: Смяна на парола чрез токен
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                newPassword:
                  type: string
              required: [token, newPassword]
      responses:
        "200":
          description: Паролата е сменена успешно
        "400":
          description: Невалиден или изтекъл токен
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [Auth]
      summary: Потвърждение на имейл чрез verification токен
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required: [token]
      responses:
        "200":
          description: Имейлът е потвърден успешно (регистрация или смяна на имейл)
        "400":
          description: Невалиден или изтекъл verification токен
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [Admin]
      summary: Списък с потребители
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Търсене по имейл
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [active, deactivated]
          description: Филтър по статус
        - in: query
          name: role
          required: false
          schema:
            type: string
            enum: [user, admin, monitoring, teacher, author]
          description: Филтър по роля
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Номер на страницата (за странициране на списъка)
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
          description: Брой потребители на страница
      responses:
        "200":
          description: Списък с потребители
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserSummary'
        "401":
          description: Неоторизиран достъп
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [Admin]
      summary: Активиране/деактивиране на потребителски акаунт
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserInput'
      responses:
        "200":
          description: Обновен потребител
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummary'
        "401":
          description: Неоторизиран достъп
        "404":
          description: Потребителят не е намерен
...
  /admin/users/stats:
    get:
      operationId: getAdminUsersStats
      tags: [Admin]
      summary: Статистика за потребители (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Aggregated stats for users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersStats'
        "401":
          description: Неоторизиран достъп

  ############################
  # Admin / Activity
  ############################
...